Move gen_keypair_file test to a more appropriate location
